% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_type.R
\name{as_assertion_factory}
\alias{as_assertion_factory}
\title{Build an assertion factory}
\usage{
as_assertion_factory(f)
}
\arguments{
\item{f}{A function whose first argument is the value to check. Additional
arguments become arguments of the assertion factory.}
}
\value{
A function with class \code{assertion_factory}.
}
\description{
\code{as_assertion_factory()} wraps a checking function and turns it into a typedr
assertion factory. The wrapped function must take the value to check as its
first argument and return the checked value, or throw an error if the value is
invalid.
}
