% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typedR.R
\name{typedR}
\alias{typedR}
\title{Animate string typing}
\usage{
typedR(text, speed = 20, theme = "mac", ratio = 0.8,
  background_col = "lightblue", width = NULL, height = NULL)
}
\arguments{
\item{text}{text to animate.}

\item{speed}{speed of the animation.}

\item{theme}{theme of the terminal. Currently only supports `mac`.}

\item{background_col}{background color of the panel outside the terminal UI.}
}
\description{
Animate string typing in a terminal like UI.
}
\examples{
typedR(text = "matrix(1:4, ncol = 2), 2, mean)",
       speed = 10, background_col = "lightblue")
}
